fix: 책장 리스트용 조회 API 파라미터 일부 수정
- readingPercent 파라미터의 타입이 다른 API와 다른 부분이 있어 이를 통일함

+++ server/src/main/java/com/codecozy/server/dto/response/DetailBookshelfResponse.java
@@ -10,5 +10,5 @@ public record DetailBookshelfResponse(
         boolean isMine,
         int totalPage,
         int readPage,
-        float readingPercent
+        int readingPercent
 ) {}
